dom/tests/mochitest/bugs/test_bug857555.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=857555
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 857555</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 857555 **/
  SimpleTest.waitForExplicitFinish();

  addLoadEvent(function() {
    is(content, $("t").contentWindow, "'content' as iframe name should work");
    is(sidebar, $("u").contentWindow, "'sidebar' as iframe name should work");
    SimpleTest.finish();
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=857555">Mozilla Bug 857555</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe name="content" id="t"></iframe>
  <iframe name="sidebar" id="u"></iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
